Create message class to represent serialized data, and create TCP and UDP clients
The templated message class is to be passed the protobuf class, so that it can be properly serialized and deserialized within the class. The class stores the data as a vector of chars. NetClient contains templated TCP and UDP clients, where the template argument is again the protobuf message type