Added the power metrics table that takes data from telemetry and displays them.
I made the power metrics table. I also edited the styling for the navigation table. changes to the dashboard page component html and css were made to get both navigation and power metrics in a column. the styling of the min actual max component was changed, some css was deleted. I used the html and css of the natural gas component to position it on the display.